Database Support+
Made style changes as well as implemented database-only options. Opted to use a database resident in memory as the default database location. Updated test file to accomodate a non-filepath db location.

+++ test/test_models.py
@@ -94,7 +94,8 @@ class TestCacheDatabase(unittest.TestCase):
         self.assertTrue(c.addr_range[0] <= addr <= c.addr_range[1])
 
     def tearDown(self):
-        os.remove(DEFAULT_DB_PATH)
+        if DEFAULT_DB_PATH != ':memory:':
+            os.remove(DEFAULT_DB_PATH)
 
 if __name__ == '__main__':
     unittest.main()
